Style header and main.
Header is fine, though padding on logo is something to think about.

Padding (especially left and right) on main has to be thought about
further for more responsiveness.

The actual form is not styled yet, that is, it isn't in two columns,
it isn't responsive, and no effects have been added for hovering,
focusing, or invalid responses.

The height of the footer on main is not occupying the rest of the space.

The button does not have styles for hover and focus.

The link for logging in does not have a style for visited.